# src/dscompanion/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dscompanion API",
    "description": "REST surface over dscompanion's 13-step interactive pipeline wizard.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Create Run",
        "description": "Creates a new run.\n\nArgs:\n    store (RunStateStore): Injected via ``get_state_store``.\n\nReturns:\n    RunCreateResponse: The new run's ``run_id``.",
        "operationId": "create_run_api_runs_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/state": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get State",
        "description": "Returns the current state of a run.\n\nArgs:\n    run (RunState): Injected via ``get_pipeline_service`` from the ``run_id`` path\n        parameter.\n\nReturns:\n    RunStateResponse: Step statuses, confirmations, audit trail, config preview.",
        "operationId": "get_state_api_runs__run_id__state_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/back": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Go Back",
        "description": "Two-phase reset to an earlier step, mirroring the UI's confirm-before-apply back\nbutton.\n\nArgs:\n    body (BackRequest): ``to_step`` and whether this call actually applies the\n        reset (``confirmed=True``) or is just asking for confirmation.\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    BackResponse: If ``body.confirmed`` is ``False``, only\n    ``requires_confirmation=True`` is set. Otherwise the reset is applied and the\n    updated state is returned.",
        "operationId": "go_back_api_runs__run_id__back_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}/steps/shap/preview": {
      "post": {
        "tags": [
          "shap"
        ],
        "summary": "Preview",
        "description": "Reports whether SHAP is available for this run, without computing anything.\n\nArgs:\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    ShapPreviewResponse: Availability \u2014 always ``True`` today.\n\nRaises:\n    HTTPException: 400, if Step 11 has not been confirmed for this run.",
        "operationId": "preview_api_runs__run_id__steps_shap_preview_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShapPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/steps/shap/run": {
      "post": {
        "tags": [
          "shap"
        ],
        "summary": "Shap Run",
        "description": "Fits ``SHAPExplainer`` on a sample of the held-out split and returns the summary.\n\nExplicit opt-in the caller must trigger \u2014 never called automatically by\n``preview``, since SHAP computation is relatively expensive.\n\nArgs:\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    ShapRunResponse: The summary chart, top-10 feature table, split\n    used, and row count.\n\nRaises:\n    HTTPException: 400, if Step 11 has not been confirmed, or no\n        held-out data is available, for this run.",
        "operationId": "shap_run_api_runs__run_id__steps_shap_run_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShapRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/steps/shap/confirm": {
      "post": {
        "tags": [
          "shap"
        ],
        "summary": "Confirm",
        "description": "Records Step 12 completion.\n\nArgs:\n    body (ShapConfirmRequest): Whether SHAP was actually run.\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    ShapConfirmResponse: Confirmation.",
        "operationId": "confirm_api_runs__run_id__steps_shap_confirm_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShapConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShapConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BackRequest": {
        "properties": {
          "to_step": {
            "type": "string",
            "enum": [
              "load_data",
              "task_target",
              "split",
              "eda",
              "feature_processing",
              "feature_selection",
              "imbalance",
              "train",
              "tuning",
              "evaluate",
              "calibration",
              "shap",
              "report"
            ],
            "title": "To Step"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "to_step"
        ],
        "title": "BackRequest",
        "description": "Request for ``POST /api/runs/{run_id}/back``.\n\nArgs:\n    to_step (StepId): The step to go back to \u2014 this step and every step\n        after it will be invalidated once confirmed.\n    confirmed (bool): ``False`` (default) for the first call \u2014 the response\n        will ask for confirmation without changing anything.  ``True`` to\n        actually apply the reset, mirroring the UI's two-phase back button."
      },
      "BackResponse": {
        "properties": {
          "requires_confirmation": {
            "type": "boolean",
            "title": "Requires Confirmation"
          },
          "state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunStateResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "requires_confirmation"
        ],
        "title": "BackResponse",
        "description": "Response for ``POST /api/runs/{run_id}/back``.\n\nArgs:\n    requires_confirmation (bool): ``True`` when this call was the first\n        phase (``confirmed=False`` in the request) \u2014 call again with\n        ``confirmed=True`` to actually apply the reset.\n    state (RunStateResponse | None): The updated state, populated only once\n        the reset has actually been applied (``requires_confirmation=False``)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunCreateResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RunCreateResponse",
        "description": "Response for ``POST /api/runs``.\n\nArgs:\n    run_id (str): The newly created run's identifier \u2014 pass this in every\n        subsequent request's path."
      },
      "RunStateResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "step_status": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "enum": [
                "load_data",
                "task_target",
                "split",
                "eda",
                "feature_processing",
                "feature_selection",
                "imbalance",
                "train",
                "tuning",
                "evaluate",
                "calibration",
                "shap",
                "report"
              ]
            },
            "type": "object",
            "title": "Step Status"
          },
          "step_confirmed": {
            "additionalProperties": {
              "type": "boolean"
            },
            "propertyNames": {
              "enum": [
                "load_data",
                "task_target",
                "split",
                "eda",
                "feature_processing",
                "feature_selection",
                "imbalance",
                "train",
                "tuning",
                "evaluate",
                "calibration",
                "shap",
                "report"
              ]
            },
            "type": "object",
            "title": "Step Confirmed"
          },
          "audit_trail": {
            "items": {
              "prefixItems": [
                {
                  "type": "string",
                  "enum": [
                    "load_data",
                    "task_target",
                    "split",
                    "eda",
                    "feature_processing",
                    "feature_selection",
                    "imbalance",
                    "train",
                    "tuning",
                    "evaluate",
                    "calibration",
                    "shap",
                    "report"
                  ]
                },
                {
                  "type": "string"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Audit Trail"
          },
          "config_preview": {
            "type": "object",
            "title": "Config Preview"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "step_status",
          "step_confirmed",
          "audit_trail",
          "config_preview"
        ],
        "title": "RunStateResponse",
        "description": "Response for ``GET /api/runs/{run_id}/state`` \u2014 the REST equivalent of\n``_interactive_flags()`` + ``_build_partial_config_dict()`` combined.\n\nArgs:\n    run_id (str): This run's identifier.\n    step_status (dict[StepId, str]): One of ``\"done\"``, ``\"current\"``,\n        ``\"pending\"``, ``\"flagged\"`` per step.\n    step_confirmed (dict[StepId, bool]): Whether each step has been\n        confirmed.\n    audit_trail (list[tuple[StepId, str]]): One plain-language entry per\n        confirmed step, in step order.\n    config_preview (dict[str, Any]): Flat dict keyed by step slug, containing\n        only the fields decided by confirmed steps so far \u2014 grows exactly in\n        step with confirmations, same philosophy as ``streamlit_app.py``'s\n        ``_build_partial_config_dict()``."
      },
      "ShapConfirmRequest": {
        "properties": {
          "shap_run": {
            "type": "boolean",
            "title": "Shap Run"
          }
        },
        "type": "object",
        "required": [
          "shap_run"
        ],
        "title": "ShapConfirmRequest",
        "description": "Request for ``POST /api/runs/{run_id}/steps/shap/confirm``.\n\nArgs:\n    shap_run (bool): Whether SHAP was actually computed (``True``) or\n        skipped by the user (``False``) \u2014 recorded in the audit trail."
      },
      "ShapConfirmResponse": {
        "properties": {
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          }
        },
        "type": "object",
        "required": [
          "confirmed"
        ],
        "title": "ShapConfirmResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/shap/confirm``.\n\nArgs:\n    confirmed (bool): Always ``True`` on success."
      },
      "ShapFeatureRow": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "feature": {
            "type": "string",
            "title": "Feature"
          },
          "mean_abs_shap": {
            "type": "number",
            "title": "Mean Abs Shap"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "feature",
          "mean_abs_shap"
        ],
        "title": "ShapFeatureRow",
        "description": "One row of Step 12's top-feature importance table.\n\nArgs:\n    rank (int): 1-based rank, most important feature first.\n    feature (str): Feature name.\n    mean_abs_shap (float): Mean absolute SHAP value across the sample."
      },
      "ShapPreviewResponse": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          }
        },
        "type": "object",
        "required": [
          "available"
        ],
        "title": "ShapPreviewResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/shap/preview``.\n\nArgs:\n    available (bool): Whether SHAP can be run for this run. Always\n        ``True`` today \u2014 exists as a forward-compatible kill switch."
      },
      "ShapRunResponse": {
        "properties": {
          "figure": {
            "type": "object",
            "title": "Figure"
          },
          "top_features": {
            "items": {
              "$ref": "#/components/schemas/ShapFeatureRow"
            },
            "type": "array",
            "title": "Top Features"
          },
          "split_used": {
            "type": "string",
            "title": "Split Used"
          },
          "n_rows": {
            "type": "integer",
            "title": "N Rows"
          }
        },
        "type": "object",
        "required": [
          "figure",
          "top_features",
          "split_used",
          "n_rows"
        ],
        "title": "ShapRunResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/shap/run``.\n\nArgs:\n    figure (dict[str, Any]): SHAP summary bar chart, JSON-safe via\n        ``fig.to_json()``.\n    top_features (list[ShapFeatureRow]): Top 10 features by mean\n        absolute SHAP value.\n    split_used (str): Held-out split SHAP was computed on (``\"val\"`` if\n        available, else ``\"test\"``, else ``\"train\"``).\n    n_rows (int): Number of rows actually sampled for the computation\n        (capped at ``settings.shap_interactive_max_rows``)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}